Add block_reader, remove file_reader from read pyi
Add the block_reader and its partner Block to the type stub
for the read module.

On the way, remove file_reader because it's not in the exported
public API, and it has an ambiguous (or at least complicated)
type for its __iter__, next, and __next__ methods; their return
value types depend upon the concrete subclass.