Apply logical type preparation when checking for best union type (#530)
* Apply logical type preparation when checking for best union type

Previously validate would pass (and it does the prepare step) but then the datum was still pre preparation when it tried to do anything further. As far as I can tell this only really matters for record types, but made more general in case there was an edge case I was missing. The provided test shows this for an ndarray record type which would fail because set(ndarray) does not work [well, it actually does for 1D arrays, but not 2+ dimensional arrays... even in the 1d case it was doing something funky that techincally worked, but only on accident]

* only do logical type conversion for records, remove superfluous import