Allow arbitrary ordering of union fields
Since the Java reference implementation no longer enforces the
requirement that the default value be the first type in a union,
drop this requirement from fastavro as well.  This restores the
ability to read data written with fastavro <1.7, while still checking
that the default is (or is convertible to) one of the union types.
Along the way, refactor the special cases for double and float so
that they do not raise exceptions.

Fixes #723