Add default value to optional argument in type stub
The function `fastavro.read.schemaless_reader` has an optional argument `reader_schema`, which by default is set to `None`. In the `_read.pyi`-type stub this argument is marked as optional, but not given a default value. This made mypy mark it as an error with the message `error: Too few arguments  [call-arg]` if `reader_schema` is not explicitly given a value, with the code working just fine.